remove "common" module
